requirements.txt
Django>=4.2.4
django-admin-datta>=1.0.10
django-cors-headers>=4.2.0
django-webpack-loader>=2.0.1
djangorestframework>=3.14.0
mysqlclient >= 2.2.0
djangorestframework-simplejwt >= 5.3.0
django-rest-swagger >= 2.2.0
drf-yasg>=1.21.7
django-filter>=23.3